Moved a lot to Core, this is so we can produce a nuget package for those that are "doing their own thing" rather than creating a WebApi
