+++ iRacingStats.Core/Constants/Config.cs
@@ -1,0 +1,7 @@
+﻿namespace iRacingStats.Core.Constants
+{
+    public class Config
+    {
+        public const int POST_DELAY = 800;
+    }
+}
