+++ iRacingStats.Core/Constants/Licenses.cs
@@ -1,0 +1,13 @@
+﻿namespace iRacingStats.Core.Constants
+{
+    public class Licenses
+    {
+        public const int ROOKIE = 1;
+        public const int D = 2;
+        public const int C = 3;
+        public const int B = 4;
+        public const int A = 5;
+        public const int PRO = 6;
+        public const int PRO_WC = 7;
+    }
+}
