+++ iRacingStats.Core/Constants/Locations.cs
@@ -1,0 +1,253 @@
+﻿namespace iRacingStats.Core.Constants
+{
+    public class Locations
+    {
+        public const string ALL = "null";
+        public const string AFGHANISTAN = "AF";
+        public const string ALAND_ISLANDS = "AX";
+        public const string ALBANIA = "AL";
+        public const string ALGERIA = "DZ";
+        public const string AMERICAN_SAMOA = "AS";
+        public const string ANDORRA = "AD";
+        public const string ANGOLA = "AO";
+        public const string ANGUILLA = "AI";
+        public const string ANTARCTICA = "AQ";
+        public const string ANTIGUA_AND_BARBUDA = "AG";
+        public const string ARGENTINA = "AR";
+        public const string ARMENIA = "AM";
+        public const string ARUBA = "AW";
+        public const string AUSTRALIA = "AU";
+        public const string AUSTRIA = "AT";
+        public const string AZERBAIJAN = "AZ";
+        public const string BAHAMAS = "BS";
+        public const string BAHRAIN = "BH";
+        public const string BANGLADESH = "BD";
+        public const string BARBADOS = "BB";
+        public const string BELARUS = "BY";
+        public const string BELGIUM = "BE";
+        public const string BELIZE = "BZ";
+        public const string BENIN = "BJ";
+        public const string BERMUDA = "BM";
+        public const string BHUTAN = "BT";
+        public const string BOLIVIA_PLURINATIONAL_STATE_OF = "BO";
+        public const string BOSNIA_AND_HERZEGOVINA = "BA";
+        public const string BOTSWANA = "BW";
+        public const string BOUVET_ISLAND = "BV";
+        public const string BRAZIL = "BR";
+        public const string BRITISH_INDIAN_OCEAN_TERRITORY = "IO";
+        public const string BRUNEI_DARUSSALAM = "BN";
+        public const string BULGARIA = "BG";
+        public const string BURKINA_FASO = "BF";
+        public const string BURUNDI = "BI";
+        public const string CAMBODIA = "KH";
+        public const string CAMEROON = "CM";
+        public const string CANADA = "CA";
+        public const string CAPE_VERDE = "CV";
+        public const string CAYMAN_ISLANDS = "KY";
+        public const string CENTRAL_AFRICAN_REPUBLIC = "CF";
+        public const string CHAD = "TD";
+        public const string CHILE = "CL";
+        public const string CHINA = "CN";
+        public const string CHRISTMAS_ISLAND = "CX";
+        public const string COCOS_KEELING_ISLANDS = "CC";
+        public const string COLOMBIA = "CO";
+        public const string COMOROS = "KM";
+        public const string CONGO = "CG";
+        public const string CONGO_THE_DEMOCRATIC_REPUBLIC_OF_THE = "CD";
+        public const string COOK_ISLANDS = "CK";
+        public const string COSTA_RICA = "CR";
+        public const string COTE_DIVOIRE = "CI";
+        public const string CROATIA = "HR";
+        public const string CUBA = "CU";
+        public const string CYPRUS = "CY";
+        public const string CZECH_REPUBLIC = "CZ";
+        public const string DENMARK = "DK";
+        public const string DJIBOUTI = "DJ";
+        public const string DOMINICA = "DM";
+        public const string DOMINICAN_REPUBLIC = "DO";
+        public const string ECUADOR = "EC";
+        public const string EGYPT = "EG";
+        public const string EL_SALVADOR = "SV";
+        public const string EQUATORIAL_GUINEA = "GQ";
+        public const string ERITREA = "ER";
+        public const string ESTONIA = "EE";
+        public const string ETHIOPIA = "ET";
+        public const string FALKLAND_ISLANDS_MALVINAS = "FK";
+        public const string FAROE_ISLANDS = "FO";
+        public const string FIJI = "FJ";
+        public const string FINLAND = "FI";
+        public const string FRANCE = "FR";
+        public const string FRENCH_GUIANA = "GF";
+        public const string FRENCH_POLYNESIA = "PF";
+        public const string FRENCH_SOUTHERN_TERRITORIES = "TF";
+        public const string GABON = "GA";
+        public const string GAMBIA = "GM";
+        public const string GEORGIA = "GE";
+        public const string GERMANY = "DE";
+        public const string GHANA = "GH";
+        public const string GIBRALTAR = "GI";
+        public const string GREECE = "GR";
+        public const string GREENLAND = "GL";
+        public const string GRENADA = "GD";
+        public const string GUADELOUPE = "GP";
+        public const string GUAM = "GU";
+        public const string GUATEMALA = "GT";
+        public const string GUERNSEY = "GG";
+        public const string GUINEA = "GN";
+        public const string GUINEA_BISSAU = "GW";
+        public const string GUYANA = "GY";
+        public const string HAITI = "HT";
+        public const string HEARD_ISLAND_AND_MCDONALD_ISLANDS = "HM";
+        public const string HOLY_SEE_VATICAN_CITY_STATE = "VA";
+        public const string HONDURAS = "HN";
+        public const string HONG_KONG = "HK";
+        public const string HUNGARY = "HU";
+        public const string ICELAND = "IS";
+        public const string INDIA = "IN";
+        public const string INDONESIA = "ID";
+        public const string IRAN_ISLAMIC_REPUBLIC_OF = "IR";
+        public const string IRAQ = "IQ";
+        public const string IRELAND = "IE";
+        public const string ISLE_OF_MAN = "IM";
+        public const string ISRAEL = "IL";
+        public const string ITALY = "IT";
+        public const string JAMAICA = "JM";
+        public const string JAPAN = "JP";
+        public const string JERSEY = "JE";
+        public const string JORDAN = "JO";
+        public const string KAZAKHSTAN = "KZ";
+        public const string KENYA = "KE";
+        public const string KIRIBATI = "KI";
+        public const string KOREA_DEMOCRATIC_PEOPLES_REPUBLIC_OF = "KP";
+        public const string KOREA_REPUBLIC_OF = "KR";
+        public const string KUWAIT = "KW";
+        public const string KYRGYZSTAN = "KG";
+        public const string LAO_PEOPLES_DEMOCRATIC_REPUBLIC = "LA";
+        public const string LATVIA = "LV";
+        public const string LEBANON = "LB";
+        public const string LESOTHO = "LS";
+        public const string LIBERIA = "LR";
+        public const string LIBYAN_ARAB_JAMAHIRIYA = "LY";
+        public const string LIECHTENSTEIN = "LI";
+        public const string LITHUANIA = "LT";
+        public const string LUXEMBOURG = "LU";
+        public const string MACAO = "MO";
+        public const string MACEDONIA_THE_FORMER_YUGOSLAV_REPUBLIC_OF = "MK";
+        public const string MADAGASCAR = "MG";
+        public const string MALAWI = "MW";
+        public const string MALAYSIA = "MY";
+        public const string MALDIVES = "MV";
+        public const string MALI = "ML";
+        public const string MALTA = "MT";
+        public const string MARSHALL_ISLANDS = "MH";
+        public const string MARTINIQUE = "MQ";
+        public const string MAURITANIA = "MR";
+        public const string MAURITIUS = "MU";
+        public const string MAYOTTE = "YT";
+        public const string MEXICO = "MX";
+        public const string MICRONESIA_FEDERATED_STATES_OF = "FM";
+        public const string MOLDOVA_REPUBLIC_OF = "MD";
+        public const string MONACO = "MC";
+        public const string MONGOLIA = "MN";
+        public const string MONTENEGRO = "ME";
+        public const string MONTSERRAT = "MS";
+        public const string MOROCCO = "MA";
+        public const string MOZAMBIQUE = "MZ";
+        public const string MYANMAR = "MM";
+        public const string NAMIBIA = "NA";
+        public const string NAURU = "NR";
+        public const string NEPAL = "NP";
+        public const string NETHERLANDS = "NL";
+        public const string NETHERLANDS_ANTILLES = "AN";
+        public const string NEW_CALEDONIA = "NC";
+        public const string NEW_ZEALAND = "NZ";
+        public const string NICARAGUA = "NI";
+        public const string NIGER = "NE";
+        public const string NIGERIA = "NG";
+        public const string NIUE = "NU";
+        public const string NORFOLK_ISLAND = "NF";
+        public const string NORTHERN_MARIANA_ISLANDS = "MP";
+        public const string NORWAY = "NO";
+        public const string OMAN = "OM";
+        public const string PAKISTAN = "PK";
+        public const string PALAU = "PW";
+        public const string PALESTINIAN_TERRITORY_OCCUPIED = "PS";
+        public const string PANAMA = "PA";
+        public const string PAPUA_NEW_GUINEA = "PG";
+        public const string PARAGUAY = "PY";
+        public const string PERU = "PE";
+        public const string PHILIPPINES = "PH";
+        public const string PITCAIRN = "PN";
+        public const string POLAND = "PL";
+        public const string PORTUGAL = "PT";
+        public const string PUERTO_RICO = "PR";
+        public const string QATAR = "QA";
+        public const string REUNION = "RE";
+        public const string ROMANIA = "RO";
+        public const string RUSSIAN_FEDERATION = "RU";
+        public const string RWANDA = "RW";
+        public const string SAINT_BARTHELEMY = "BL";
+        public const string SAINT_HELENA_ASCENSION_AND_TRISTAN_DA_CUNHA = "SH";
+        public const string SAINT_KITTS_AND_NEVIS = "KN";
+        public const string SAINT_LUCIA = "LC";
+        public const string SAINT_MARTIN_FRENCH_PART = "MF";
+        public const string SAINT_PIERRE_AND_MIQUELON = "PM";
+        public const string SAINT_VINCENT_AND_THE_GRENADINES = "VC";
+        public const string SAMOA = "WS";
+        public const string SAN_MARINO = "SM";
+        public const string SAO_TOME_AND_PRINCIPE = "ST";
+        public const string SAUDI_ARABIA = "SA";
+        public const string SENEGAL = "SN";
+        public const string SERBIA = "RS";
+        public const string SEYCHELLES = "SC";
+        public const string SIERRA_LEONE = "SL";
+        public const string SINGAPORE = "SG";
+        public const string SLOVAKIA = "SK";
+        public const string SLOVENIA = "SI";
+        public const string SOLOMON_ISLANDS = "SB";
+        public const string SOMALIA = "SO";
+        public const string SOUTH_AFRICA = "ZA";
+        public const string SOUTH_GEORGIA_AND_THE_SOUTH_SANDWICH_ISLANDS = "GS";
+        public const string SPAIN = "ES";
+        public const string SRI_LANKA = "LK";
+        public const string SUDAN = "SD";
+        public const string SURINAME = "SR";
+        public const string SVALBARD_AND_JAN_MAYEN = "SJ";
+        public const string SWAZILAND = "SZ";
+        public const string SWEDEN = "SE";
+        public const string SWITZERLAND = "CH";
+        public const string SYRIAN_ARAB_REPUBLIC = "SY";
+        public const string TAIWAN_PROVINCE_OF_CHINA = "TW";
+        public const string TAJIKISTAN = "TJ";
+        public const string TANZANIA_UNITED_REPUBLIC_OF = "TZ";
+        public const string THAILAND = "TH";
+        public const string TIMOR_LESTE = "TL";
+        public const string TOGO = "TG";
+        public const string TOKELAU = "TK";
+        public const string TONGA = "TO";
+        public const string TRINIDAD_AND_TOBAGO = "TT";
+        public const string TUNISIA = "TN";
+        public const string TURKEY = "TR";
+        public const string TURKMENISTAN = "TM";
+        public const string TURKS_AND_CAICOS_ISLANDS = "TC";
+        public const string TUVALU = "TV";
+        public const string UGANDA = "UG";
+        public const string UKRAINE = "UA";
+        public const string UNITED_ARAB_EMIRATES = "AE";
+        public const string UNITED_KINGDOM = "GB";
+        public const string UNITED_STATES = "US";
+        public const string UNITED_STATES_MINOR_OUTLYING_ISLANDS = "UM";
+        public const string URUGUAY = "UY";
+        public const string UZBEKISTAN = "UZ";
+        public const string VANUATU = "VU";
+        public const string VENEZUELA_BOLIVARIAN_REPUBLIC_OF = "VE";
+        public const string VIET_NAM = "VN";
+        public const string VIRGIN_ISLANDS_BRITISH = "VG";
+        public const string VIRGIN_ISLANDS_US = "VI";
+        public const string WALLIS_AND_FUTUNA = "WF";
+        public const string WESTERN_SAHARA = "EH";
+        public const string YEMEN = "YE";
+        public const string ZAMBIA = "ZM";
+        public const string ZIMBABWE = "ZW";
+    }
+}
