+++ iRacingStats.Core/Options/User.cs
@@ -1,0 +1,8 @@
+﻿namespace iRacingStats.Core.Options
+{
+    public class User
+    {
+        public string Username { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+    }
+}
