--- iRacingStatsAPI/Constants/Config.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace iRacingStatsAPI.Constants
-{
-    public class Config
-    {
-        public const int POST_DELAY = 800;
-    }
-}
