--- iRacingStatsAPI/GlobalSuppressions.cs
@@ -1,8 +1,0 @@
-﻿// This file is used by Code Analysis to maintain SuppressMessage
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given
-// a specific target and scoped to a namespace, type, member, etc.
-
-using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "No control over endpoint datamodel naming.", Scope = "namespaceanddescendants", Target = "~N:iRacingStatsAPI.Models")]
