--- iRacingStatsAPI/Options/User.cs
@@ -1,8 +1,0 @@
-﻿namespace iRacingStatsAPI.Options
-{
-    public class User
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
