+++ iRacingStatsAPI/Program.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace iRacingStatsAPI
+namespace iRacingStats.Api
 {
     public class Program
     {
