+++ iRacingStatsAPI/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "AllowedHosts": "*",
-  "iRacingStatsAPI": {
+  "iRacingStats.Core": {
     "User": {
       "Username": "",
       "Password": ""
