--- iRacingStatsAPI/iRacingStatsAPI.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-    <UserSecretsId>025188c2-c090-4706-ba97-cc90ff1df423</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Enums.NET" Version="4.0.0" />
-    <PackageReference Include="Macross.Json.Extensions" Version="2.1.0-beta5" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
-  </ItemGroup>
-
-</Project>
